+++ examples/01 - hello_world/hello.py
@@ -1,4 +1,12 @@
+import os
+import sys
+
+sys.path.insert(0, os.path.join(os.path.dirname(os.path.abspath(__file__)), '..'))
+from common import WINDOW_SIZE, use_shared_assets
+
 import glue
+
+use_shared_assets()
 
 # Set web files folder
 glue.init('web')
@@ -10,4 +18,4 @@ def say_hello_py(x):
 say_hello_py('Python World!')
 glue.say_hello_js('Python World!')   # Call a Javascript function
 
-glue.start('hello.html', size=(300, 200))  # Start
+glue.start('hello.html', size=WINDOW_SIZE)