+++ examples/06 - input/script.py
@@ -1,5 +1,12 @@
+import os
+import sys
+
+sys.path.insert(0, os.path.join(os.path.dirname(os.path.abspath(__file__)), '..'))
+from common import WINDOW_SIZE, use_shared_assets
+
 import glue
 
+use_shared_assets()
 glue.init('web')                     # Give folder containing web files
 
 @glue.expose                         # Expose this function to Javascript
@@ -8,4 +15,4 @@ def handleinput(x):
 
 glue.say_hello_js('connected!')   # Call a Javascript function
 
-glue.start('main.html', size=(500, 200))    # Start
+glue.start('main.html', size=WINDOW_SIZE)