+++ examples/07 - jinja_templates/hello.py
@@ -1,7 +1,13 @@
+import os
+import sys
 import random
+
+sys.path.insert(0, os.path.join(os.path.dirname(os.path.abspath(__file__)), '..'))
+from common import WINDOW_SIZE, use_shared_assets
 
 import glue
 
+use_shared_assets()
 glue.init('web')                     # Give folder containing web files
 
 @glue.expose
@@ -15,4 +21,4 @@ def say_hello_py(x):
 say_hello_py('Python World!')
 glue.say_hello_js('Python World!')   # Call a Javascript function
 
-glue.start('templates/hello.html', size=(300, 200), jinja_templates='templates')    # Start
+glue.start('templates/hello.html', size=WINDOW_SIZE, jinja_templates='templates')