+++ examples/10 - custom_app_routes/custom_app.py
@@ -1,12 +1,20 @@
+import os
+import sys
+
+sys.path.insert(0, os.path.join(os.path.dirname(os.path.abspath(__file__)), '..'))
+from common import WINDOW_SIZE, use_shared_assets
+
 import glue
 import bottle
 # from beaker.middleware import SessionMiddleware
 
 app = bottle.Bottle()
+
 @app.route('/custom')
 def custom_route():
     return 'Hello, World!'
 
+use_shared_assets(app)
 glue.init('web')
 
 # need to manually add glue routes if we are wrapping our Bottle instance with middleware
@@ -14,4 +22,4 @@ glue.init('web')
 # middleware = SessionMiddleware(app)
 # glue.start('index.html', app=middleware)
 
-glue.start('index.html', app=app)
+glue.start('index.html', app=app, size=WINDOW_SIZE)